aegis_brain/requirements.txt
requests==2.31.0
google-genai
docker==7.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

pymoo>=0.6.0

pandas>=2.0.0
numpy>=1.24.0

pandas_ta>=0.3.14
pyarrow>=12.0.0
